+++ Kotlin/Advanced/src/test/kotlin/MyTest.kt
@@ -1,0 +1,11 @@
+import org.junit.Test
+
+class MyTest {
+
+    @Test
+    fun testMain() {
+        println("testing")
+
+        main(arrayOf())
+    }
+}